Split module loading `IPythonModuleLoader` and `DefaultPythonModuleLoader`
